refactor(server): extract PublicError into errors module
