+++ web_monitoring_diff/server/server.py
@@ -21,6 +21,7 @@ import tornado.web
 import traceback
 import web_monitoring_diff
 from .mock_http import MockResponse
+from .errors import PublicError
 from .. import basic_diffs, html_render_diff, html_links_diff
 from ..exceptions import UndiffableContentError, UndecodableContentError
 from ..utils import shutdown_executor_in_loop, Signal
@@ -132,42 +133,6 @@ tornado.httpclient.AsyncHTTPClient.configure(HTTP_CLIENT,
 
 def get_http_client():
     return tornado.httpclient.AsyncHTTPClient()
-
-
-class PublicError(tornado.web.HTTPError):
-    """
-    Customized version of Tornado's HTTP error designed for reporting publicly
-    visible error messages. Please always raise this instead of calling
-    `send_error()` directly, since it lets you attach a user-visible
-    explanation of what went wrong.
-
-    Parameters
-    ----------
-    status_code : int, default: 500
-        Status code for the response.
-    public_message : str, optional
-        Textual description of the error. This will be publicly visible in
-        production mode, unlike `log_message`.
-    log_message : str, optional
-        Error message written to logs and to error tracking service. Will be
-        included in the HTTP response only in debug mode. Same as the
-        `log_message` parameter to `tornado.web.HTTPError`, but with no
-        interpolation.
-    extra : dict, optional
-        Dict of additional keys and values to include in the error response.
-    """
-    def __init__(self, status_code=500, public_message=None, log_message=None,
-                 extra=None, **kwargs):
-        self.extra = extra or {}
-
-        if public_message is not None:
-            if 'error' not in self.extra:
-                self.extra['error'] = public_message
-
-            if log_message is None:
-                log_message = public_message
-
-        super().__init__(status_code, log_message, **kwargs)
 
 
 DEBUG_MODE = os.environ.get('DIFFING_SERVER_DEBUG', 'False').strip().lower() == 'true'
